Convert ItemView to functional component
